NPCBots: fix return types for compatibility inliners

+++ src/server/game/Entities/Unit/Unit.h
@@ -1433,9 +1433,9 @@ public:
     //npcbot: compatibility accessors
     [[nodiscard]] inline uint8 GetLevel() const { return getLevel(); }
     [[nodiscard]] inline uint8 GetRace(bool original = false) const { return getRace(original); }
-    [[nodiscard]] inline uint8 GetRaceMask() const { return getRaceMask(); }
+    [[nodiscard]] inline uint32 GetRaceMask() const { return getRaceMask(); }
     [[nodiscard]] inline uint8 GetClass() const { return getClass(); }
-    [[nodiscard]] inline uint8 GetClassMask() const { return getClassMask(); }
+    [[nodiscard]] inline uint32 GetClassMask() const { return getClassMask(); }
     [[nodiscard]] inline uint8 GetGender() const { return getGender(); }
     inline void SetPowerType(Powers power) { setPowerType(power); }
     [[nodiscard]] inline Powers GetPowerType() const { return getPowerType(); }
